Implemented module structure Had to rename from `robust-genetics` because of the `-` which Python doesn't support in module names. Chose `alphargs` b/c we're using `alphasim` output data
